DATAFLOW-APACHE_BEAM/requirements.txt
annotated-types==0.7.0
apache-beam==2.57.0
asttokens==2.4.1
attrs==23.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
cloudpickle==2.2.1
comm==0.2.2
crcmod==1.7
debugpy==1.8.2
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.1.1
dnspython==2.6.1
docopt==0.6.2
docstring_parser==0.16
executing==2.0.1
facets-overview==1.1.1
fastavro==1.9.4
fasteners==0.19
fastjsonschema==2.20.0
google-api-core==2.19.1
google-apitools==0.5.31
google-auth==2.31.0
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.57.0
google-cloud-bigquery==3.25.0
google-cloud-bigquery-storage==2.25.0
google-cloud-bigtable==2.24.0
google-cloud-core==2.4.1
google-cloud-dataproc==5.10.0
google-cloud-datastore==2.19.0
google-cloud-dlp==3.18.0
google-cloud-language==2.13.3
google-cloud-pubsub==2.21.5
google-cloud-pubsublite==1.10.0
google-cloud-recommendations-ai==0.10.10
google-cloud-resource-manager==1.12.3
google-cloud-spanner==3.47.0
google-cloud-storage==2.17.0
google-cloud-videointelligence==2.13.3
google-cloud-vision==3.7.2
google-crc32c==1.5.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.2
grpc-google-iam-v1==0.13.1
grpc-interceptor==0.15.4
grpcio==1.64.1
grpcio-status==1.62.2
hdfs==2.7.3
httplib2==0.22.0
idna==3.7
ipykernel==6.29.5
ipython==8.26.0
ipywidgets==8.1.3
jedi==0.19.1
Jinja2==3.1.4
jsonpickle==3.2.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_client==8.2.0
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
jupyterlab_widgets==3.0.11
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mistune==3.0.2
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
numpy==1.26.4
oauth2client==4.1.3
objsize==0.7.0
orjson==3.10.5
overrides==7.7.0
packaging==24.1
pandas==2.1.4
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==4.25.3
psutil==6.0.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==16.1.0
pyarrow-hotfix==0.6
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.8.0
pydantic_core==2.20.0
pydot==1.4.2
Pygments==2.18.0
pymongo==4.8.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
pyzmq==26.0.3
redis==5.0.7
referencing==0.35.1
regex==2024.5.15
requests==2.31.0
rpds-py==0.18.1
rsa==4.9
shapely==2.0.4
six==1.16.0
soupsieve==2.5
sqlparse==0.5.0
stack-data==0.6.3
timeloop==1.0.2
tinycss2==1.3.0
tornado==6.4.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
wcwidth==0.2.13
webencodings==0.5.1
widgetsnbextension==4.0.11
zstandard==0.22.0
